capstone/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

# ────── DB ──────
sqlalchemy==1.4.46
psycopg2-binary>=2.9.9,<3

# ────── Auth / 보안 ──────
passlib==1.7.4
PyJWT==2.6.0
python-jose[cryptography]==3.3.0

python-multipart==0.0.7
python-dotenv==0.21.0

# ────── Google Gemini / LLM ──────
# langchain-google-genai 1.x는 google-generativeai<0.8 과 호환 (0.2.x LangChain 스택과 맞춤)
google-generativeai>=0.7.0,<0.8.0
httpx>=0.27.0

# ────── PDF 요약 ──────
pdfplumber==0.9.0

# ────── Google OAuth / Calendar ──────
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# ────── 이미지 처리 ──────
Pillow==10.3.0           # 썸네일·WebP 저장
requests==2.31.0         # 이미지 URL 다운로드

# ────── LangChain ──────
langchain==0.2.17
langchain-google-genai==1.0.10  # 2.x+ 는 langchain-core 1.x 필요 — 0.2 스택에 고정
langchain-core==0.2.43